step2 - Inventory test success
